backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
langchain
langchain-openai
faiss-cpu
sqlite-utils
python-dotenv